Squashed 'Template2D/Assets/Modules/' changes from 1d405cf..add2caf
add2caf Summary:    * More work on the state machine editor V1.  Now I have arrays and stuff working somewhat nicely.  Some of it needs to be better (like more foldouts and stuff) but I'm getting there.
a43df09 Summary:    * Some more work on the state machines.  Better interface as well as some helper functions for rendering arrays in editors.  Want to find a generic way of doing that array stuff...  We'll see what I can come up with.                        * Made it so now when you generate a machine it prompts you for the folder that you want to generate the state machine in.
c612ef5 Summary:    * Fixed a couple bugs in the T4s                        * More work on the API that will be used in Unity to generate the state machines.                        * Test editor (that I might even turn into the real editor at some point) for state machine generation is in.  It can save and load right now.
8c71358 Summary:    * Decided that the next thing I wanted to work on was basic AI and for that I want some kind of state machine so that's what I started to make is a T4 that will generate enum based state machine with transitions.
49c2730 Summary:    * Added a simple, type-safe event messaging system.  Will be the major event system for all games.                        * Converted FlowManager to use the new event system instead of having it's own functions for firing a new scene                        * Cleaned up all the warnings that were all over the project.                        * Added a template scene for the application initializer into the modules.                        * Re-ordered the scripts so that they are in a folder called Scripts to make it so we can put other things in that modules folder.
0b871ff Summary:    * Re-organized a bunch of files to be in their proper namespaces.                       * Turned all the Debug.Logs to Services.Logger.Logs so that they get processed the way we want them to be.
db7435c Summary:  * Decided that instead of using a interface, I'll do it with a actual base class so that the shared functionality is all in one place.                      * Categories WORK!  You can now toggle on specific categories if you want to limit what shows up in the log

git-subtree-dir: Template2D/Assets/Modules
git-subtree-split: add2caf76bdaecf66cfbfab5e1386c3453733626